数据导入导出修改，导入修改股权权重，修改批量导出为户序号排序导出

+++ src/main/java/zyxhj/jiti/domain/ORG.java
@@ -115,7 +115,7 @@ public class ORG {
 	 * 股份份数总数
 	 */
 	@RDSAnnField(column = RDSAnnField.INTEGER)
-	public Integer shareAmount;
+	public Double shareAmount;
 	/**
 	 * 资源股
 	 */
